Refactor, Bugfix, Breaking changes (apps/pet-shop): Rename `ps-products-storage.ts`
